Add get() etc. in parent class.
This is a different approach to adding get() etc. to parse tree classes.
They are defined in base classes, TupleClass and UnionClass.

It adds get<I>() for constant I for tuples, and get_if() (both type-based
and index-based) for unions. This reduces the use of preprocessor macros.

One negative of this approach is that classes that wrap tuples and
variants can't use nested classes. See StarSize, for example.

This still doesn't handle visit() with non-void result. That is used
several times in parse-tree.cc.

If we take this approach we might want to add WrapperClass<T> with a
get() member function for consistency.