Add Visit() helper to replace std::visit+visitors.
`Visit(variant, lambda1, lambda2, ...)` is a utility function that is
equivalent to `std::visit(visitors{lambda1, lambda2, ...}, variant)`.

The advantages of using it are:
- clang-format is less likely to make a mess formatting them
- it saves 4 spaces of indentation (or more) and 2 or 3 lines when formatted
- the variant argument is at the beginning instead of lost after the lambdas

The disadvantage is that it is not standard like `std::visit` is. But in
order to use `std::visit` we have to use the non-standard `visitors` class,
so it's not fully standard anyway.