OpenMP clause validity check framework (PARALLEL as example)
1. Added an OmpClauseChecker to handle all the OpenMP nodes

2. Once we see a directive, keep its source for potential error msg and initialize
   the set of the legal clauses

3. Once we see a OmpClause (conceptual node), keep its source for potential error msg

4. Once we see a actual clause, check against the current set and throw the error msgs if anychar

Note that source provenance is kept for Directives and Clauses during parsing.

This commit takes "$OMP PARALLEL" as an example to show the framework which will
be expanded in the future. A simple test is created to check some illegal clauses.
All the clauses will be checked through different directives eventually.