Some cleanup in characteristics.{h,cc}
Change `DummyArgument` from a typedef to a struct so that `Characterize`,
`IsOptional`, `SetOptional`, and `Dump` can be member functions. This is
to help make future compatibility checking code more uniform.

Reorder declarations in `TypeAndShape` to be consistent with the other
classes. Reorder functions in `characteristics.cc` to match their order
in `characteristics.h`.