Move IsKindCompatible into DerivedTypeSpec
Make static function IsKindCompatible a member function of
DerivedTypeSpec so that it can replace one of the overloadings
of `HaveCompatibleKindParameters` in `check-allocate.cc`.

Change it to use `OrderParameterDeclarations` as
`HaveCompatibleKindParameters` did. Otherwise it doesn't compare
kind parameters from the parent type. Add a new test case that
exercises that change.

It needs `PointeeComparison`, so move that to common.