Add new flang driver main() entrypoint
Note' : This a working in progress to have a flang driver for 18. More
changes will be needed to fully support the new compiler.

Note'': Lit regression tests are added along with related CMakeLists
infrastructure, this may be the first patch to do so.

This work implements the "driver" part of an RFC sent to cfe-dev in [1].

A new entrypoint adds a binary called bin/flang.
The old flang.sh is removed from the install to make way for this.
bin/flang is implemented in terms of clang::Driver, defaulting to
the `--driver-mode=flang` mode recently added to clang in [1]. This
ensures that the driver runs in flang mode regardless of the name of
the binary inferred from `argv[0]`.
"flang -fc1" will be implemented as part of bin/flang in later patches.

If compiled against an old libclangDriver which is not flang-mode
aware, the driver fails with:
  "unsupported argument 'flang' to option '--driver-mode='"

Regression tests are added in test/driver, which mirror those in the
clang/test/Driver directory. The primary difference is that the clang
tests run `clang --driver-mode=flang <input>`, whereas these tests
run `flang <input>`.

In order to add regression tests, the relevant CMakeLists
infrastructure is included.

The configuration for the tests are in lit.cfg files. These rely on the
presence of llvm-lit and FileCheck, as long as F18 is not a part of the
monorepo it needs requires -DLLVM_EXTERNAL_LIT to be defined at cmake
configuration time to point to /bin/llvm-lit from an LLVM build.

[1] "RFC: Adding a fortran mode to the clang driver for flang"
    http://lists.llvm.org/pipermail/cfe-dev/2019-June/062669.html
[2] [clang][driver] Add basic --driver-mode=flang support for fortran
    https://reviews.llvm.org/D63607

Signed-off-by: Caroline Concatto <caroline.concatto@arm.com>