Regression tests configuration for f18 repository
The configuration for the tests are in lit.* files. These rely on the
presence of llvm-lit and FileCheck, as long as F18 is not a part of the
monorepo it needs requires -DLLVM_EXTERNAL_LIT to be defined at cmake
configuration time to point to /bin/llvm-lit from an LLVM build.

This patch:
 1) Changes CMakeLists.txt to:
  * Add the new folder with the regression tests configurations
  * Add lit configuration in test-lit/CMakeLists.txt
 2) Adds the test-lit folder with the bare minimum configuration of
    lit.cfg.py and lit.site.cfg.py.in to make a simple version lit test to work

Signed-off-by: Caroline Concatto <caroline.concatto@arm.com>